Generalise guest rename logic into SetGuestUsername component for reuse
